PROJ-48 refactor: move api from ForgotPassword to shared/api/passwordRecoveryApi
