+++ app/store.ts
@@ -1,7 +1,7 @@
 import { appReducer } from '@/entities'
 import { expiredEmailLinkApi, signUpApi } from '@/features'
-import { passwordRecoveryApi } from '@/features/forgotPassword/api'
 import { setPasswordApi } from '@/features/newPassword/model/api'
+import { passwordRecoveryApi } from '@/shared'
 import { configureStore } from '@reduxjs/toolkit'
 
 export const store = configureStore({
