+++ features/expiredEmailLink/api/types.ts
@@ -1,6 +1,3 @@
 export type ResendConfirmationCodeArgs = {
   email: string
 }
-export type ValidEmailArgs = {
-  email: string
-}
