+++ shared/api/index.ts
@@ -1,1 +1,2 @@
 export * from './config'
+export * from './passwordRecoveryApi'
